refactor: move shared gql types to core module
