+++ packages/graphql-api/src/modules/accounts/index.ts
@@ -8,13 +8,13 @@ import getSchemaDef from './schema/schema-def';
 import { Query } from './resolvers/query';
 import { Mutation } from './resolvers/mutation';
 import { User as UserResolvers } from './resolvers/user';
-import { LoginResult as LoginResultResolvers } from './resolvers/loginResult';
 import { User } from '@accounts/types';
 import { AccountsPasswordModule } from '../accounts-password';
 import { AuthenticatedDirective } from '../../utils/authenticated-directive';
 import { context } from '../../utils';
 import AccountsPassword from '@accounts/password';
 import { mergeTypeDefs } from 'graphql-toolkit';
+import { CoreAccountsModule } from '../core';
 
 export interface AccountsRequest {
   req: IncomingMessage;
@@ -50,7 +50,7 @@ export const AccountsModule: GraphQLModule<
   typeDefs: ({ config }) =>
     mergeTypeDefs(
       [
-        TypesTypeDefs(config),
+        TypesTypeDefs,
         getQueryTypeDefs(config),
         getMutationTypeDefs(config),
         ...getSchemaDef(config),
@@ -64,18 +64,21 @@ export const AccountsModule: GraphQLModule<
       [config.rootQueryName || 'Query']: Query,
       [config.rootMutationName || 'Mutation']: Mutation,
       User: UserResolvers,
-      LoginResult: LoginResultResolvers,
     } as any),
   // If necessary, import AccountsPasswordModule together with this module
-  imports: ({ config }) =>
-    config.accountsServer.getServices().password
+  imports: ({ config }) => [
+    CoreAccountsModule.forRoot({
+      userAsInterface: config.userAsInterface,
+    }),
+    ...(config.accountsServer.getServices().password
       ? [
           AccountsPasswordModule.forRoot({
             accountsPassword: config.accountsServer.getServices().password as AccountsPassword,
             ...config,
           }),
         ]
-      : [],
+      : []),
+  ],
   providers: ({ config }) => [
     {
       provide: AccountsServer,
