--- packages/graphql-api/src/modules/accounts/resolvers/loginResult.ts
@@ -1,1 +1,0 @@
-export const LoginResult = {};
