+++ packages/graphql-api/src/modules/accounts/schema/types.ts
@@ -1,35 +1,12 @@
 import gql from 'graphql-tag';
-import { AccountsModuleConfig } from '..';
 
-export default ({ userAsInterface }: AccountsModuleConfig) => gql`
+export default gql`
   directive @auth on FIELD_DEFINITION | OBJECT
-
-  type Tokens {
-    refreshToken: String
-    accessToken: String
-  }
-
-  type LoginResult {
-    sessionId: String
-    tokens: Tokens
-    user: User
-  }
 
   type ImpersonateReturn {
     authorized: Boolean
     tokens: Tokens
     user: User
-  }
-
-  type EmailRecord {
-    address: String
-    verified: Boolean
-  }
-
-  ${userAsInterface ? 'interface' : 'type'} User {
-    id: ID!
-    emails: [EmailRecord!]
-    username: String
   }
 
   input UserInput {
